pyproject.toml
[project]
name = "calcounter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6",
    "langchain>=1.2.10",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
